+++ MangoAPI.Infrastructure/Database/Configurations/ChatEntityConfiguration.cs
@@ -12,9 +12,9 @@ public class ChatEntityConfiguration : IEntityTypeConfiguration<ChatEntity>
     public void Configure(EntityTypeBuilder<ChatEntity> builder)
     {
         builder.ToTable(nameof(ChatEntity), MangoDbContext.DefaultSchema);
-        
+
         builder.HasKey(x => x.Id);
-        
+
         builder.Property(x => x.CommunityType).IsRequired();
         builder.Property(x => x.CreatedAt).IsRequired();
         builder.Property(x => x.Title).IsRequired();
@@ -28,7 +28,7 @@ public class ChatEntityConfiguration : IEntityTypeConfiguration<ChatEntity>
             {
                 Id = SeedDataConstants.WsbId,
                 Title = "WSB",
-                CommunityType = (int) CommunityType.PublicChannel,
+                CommunityType = CommunityType.PublicChannel,
                 CreatedAt = DateTime.UtcNow,
                 UpdatedAt = DateTime.UtcNow,
                 Description = "WSB Public Group",
@@ -42,7 +42,7 @@ public class ChatEntityConfiguration : IEntityTypeConfiguration<ChatEntity>
             {
                 Id = SeedDataConstants.ExtremeCodeMainId,
                 Title = "Extreme Code Main",
-                CommunityType = (int) CommunityType.PublicChannel,
+                CommunityType = CommunityType.PublicChannel,
                 Description = "Extreme Code Main Public Group",
                 CreatedAt = DateTime.UtcNow,
                 UpdatedAt = DateTime.UtcNow,
@@ -56,7 +56,7 @@ public class ChatEntityConfiguration : IEntityTypeConfiguration<ChatEntity>
             {
                 Id = SeedDataConstants.ExtremeCodeFloodId,
                 Title = "Extreme Code Flood",
-                CommunityType = (int) CommunityType.PublicChannel,
+                CommunityType = CommunityType.PublicChannel,
                 Description = "Extreme Code Flood Public Group",
                 CreatedAt = DateTime.UtcNow,
                 UpdatedAt = DateTime.UtcNow,
@@ -70,7 +70,7 @@ public class ChatEntityConfiguration : IEntityTypeConfiguration<ChatEntity>
             {
                 Id = SeedDataConstants.ExtremeCodeCppId,
                 Title = "Extreme Code C++",
-                CommunityType = (int) CommunityType.PublicChannel,
+                CommunityType = CommunityType.PublicChannel,
                 Description = "Extreme Code C++ Public Group",
                 CreatedAt = DateTime.UtcNow,
                 UpdatedAt = DateTime.UtcNow,
@@ -84,7 +84,7 @@ public class ChatEntityConfiguration : IEntityTypeConfiguration<ChatEntity>
             {
                 Id = SeedDataConstants.ExtremeCodeDotnetId,
                 Title = "Extreme Code .NET",
-                CommunityType = (int) CommunityType.PublicChannel,
+                CommunityType = CommunityType.PublicChannel,
                 Description = "Extreme Code .NET Public Group",
                 CreatedAt = DateTime.UtcNow,
                 UpdatedAt = DateTime.UtcNow,
@@ -101,7 +101,7 @@ public class ChatEntityConfiguration : IEntityTypeConfiguration<ChatEntity>
                 Description = "Direct chat between Khachatur Khachatryan and razumovsky r",
                 CreatedAt = DateTime.UtcNow,
                 UpdatedAt = DateTime.UtcNow,
-                CommunityType = (int) CommunityType.DirectChat,
+                CommunityType = CommunityType.DirectChat,
                 MembersCount = 2,
                 LastMessageAuthor = "razumovsky r",
                 LastMessageText = "Hello world!",
@@ -114,7 +114,7 @@ public class ChatEntityConfiguration : IEntityTypeConfiguration<ChatEntity>
                 Description = "Direct chat between Мусяка Колбасяка and razumovsky r",
                 CreatedAt = DateTime.UtcNow,
                 UpdatedAt = DateTime.UtcNow,
-                CommunityType = (int) CommunityType.DirectChat,
+                CommunityType = CommunityType.DirectChat,
                 MembersCount = 2,
                 LastMessageAuthor = "razumovsky r",
                 LastMessageText = "Hello world!",
@@ -127,7 +127,7 @@ public class ChatEntityConfiguration : IEntityTypeConfiguration<ChatEntity>
                 Description = "Direct chat between Amelit and razumovsky r",
                 CreatedAt = DateTime.UtcNow,
                 UpdatedAt = DateTime.UtcNow,
-                CommunityType = (int) CommunityType.DirectChat,
+                CommunityType = CommunityType.DirectChat,
                 MembersCount = 2,
                 LastMessageAuthor = "razumovsky r",
                 LastMessageText = "Hello world!",
@@ -140,7 +140,7 @@ public class ChatEntityConfiguration : IEntityTypeConfiguration<ChatEntity>
                 Description = "Direct chat between Khachatur Khachatryan and Мусяка Колбасяка",
                 CreatedAt = DateTime.UtcNow,
                 UpdatedAt = DateTime.UtcNow,
-                CommunityType = (int) CommunityType.DirectChat,
+                CommunityType = CommunityType.DirectChat,
                 MembersCount = 2,
                 LastMessageAuthor = "Khachatur Khachatryan",
                 LastMessageText = "Hello world!",
@@ -153,7 +153,7 @@ public class ChatEntityConfiguration : IEntityTypeConfiguration<ChatEntity>
                 Description = "Direct chat between Petro Kolosov and Szymon Murawski",
                 CreatedAt = DateTime.UtcNow,
                 UpdatedAt = DateTime.UtcNow,
-                CommunityType = (int) CommunityType.DirectChat,
+                CommunityType = CommunityType.DirectChat,
                 MembersCount = 2,
                 LastMessageAuthor = "Petro Kolosov",
                 LastMessageText = "Hello world!",
